New tap and various test tweaks
This also splits the scope-in-config into two, since init() doesn't
reload the package.json on each call, so it can really only be done once
per process.
