+++ test/name-spaces.js
@@ -4,8 +4,10 @@ var rimraf = require('rimraf')
 var common = require('./lib/common')
 
 test('spaces', function (t) {
+  rimraf.sync(__dirname + '/package.json')
   init(__dirname, '', {}, function (er, data) {
-    t.ok(!er, 'should not error')
+    if (er)
+      throw er
     var wanted = {
       name: 'the-name',
       version: '1.0.0',
@@ -15,7 +17,8 @@ test('spaces', function (t) {
       author: '',
       main: 'basic.js'
     }
-    t.same(data, wanted)
+    console.log('')
+    t.has(data, wanted)
     t.end()
   })
   common.drive([
