+++ test/scope-in-config.js
@@ -21,22 +21,8 @@ tap.test('--yes with scope', function (t) {
   init(__dirname, __dirname, { yes: 'yes', scope: '@scoped' }, function (er, data) {
     if (er) throw er
 
-    t.same(EXPECT, data)
-    t.end()
-  })
-})
-
-var json = {
-  name: '@already/scoped',
-  version: '1.0.0'
-}
-
-tap.test('with existing package.json', function (t) {
-  fs.writeFileSync(path.join(__dirname, 'package.json'), JSON.stringify(json, null, 2))
-  init(__dirname, __dirname, { yes: 'yes', scope: '@still' }, function (er, data) {
-    if (er) throw er
-
-    t.equal(data.name, '@still/scoped', 'new scope is added, basic name is kept')
+    console.log('')
+    t.has(data, EXPECT)
     t.end()
   })
 })
